requirements.txt
django
djangorestframework
google-api-python-client
django_filter